refactor the dodgy stuff out as extension fns
