+++ src/main/kotlin/dartzee/screen/game/golf/GamePanelGolf.kt
@@ -6,6 +6,7 @@ import dartzee.achievements.ACHIEVEMENT_REF_GOLF_POINTS_RISKED
 import dartzee.achievements.retrieveAchievementForDetail
 import dartzee.ai.AbstractDartsModel
 import dartzee.core.obj.HashMapList
+import dartzee.core.util.doGolfMiss
 import dartzee.db.AchievementEntity
 import dartzee.db.GameEntity
 import dartzee.db.ParticipantEntity
