+++ src/main/kotlin/dartzee/screen/game/rtc/GamePanelRoundTheClock.kt
@@ -7,6 +7,8 @@ import dartzee.ai.AbstractDartsModel
 import dartzee.core.obj.HashMapCount
 import dartzee.core.obj.HashMapList
 import dartzee.core.util.Debug
+import dartzee.core.util.doBadLuck
+import dartzee.core.util.doForsyth
 import dartzee.db.AchievementEntity
 import dartzee.db.GameEntity
 import dartzee.screen.game.AbstractDartsGameScreen
