+++ src/main/kotlin/dartzee/screen/game/x01/GamePanelX01.kt
@@ -6,6 +6,9 @@ import dartzee.achievements.*
 import dartzee.ai.AbstractDartsModel
 import dartzee.core.obj.HashMapCount
 import dartzee.core.obj.HashMapList
+import dartzee.core.util.doBadLuck
+import dartzee.core.util.doFawlty
+import dartzee.core.util.playDodgySound
 import dartzee.db.AchievementEntity
 import dartzee.db.GameEntity
 import dartzee.db.X01FinishEntity
